backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Celery==5.3.4
redis==5.0.1
requests==2.31.0
PyPDF2==3.0.1
lxml==4.9.3
beautifulsoup4==4.12.2
python-magic==0.4.27
Pillow==10.0.1
pytesseract==0.3.10
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==2.3.7